uw_backend/requirements.txt
asn1crypto==1.5.1
blinker==1.9.0
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cryptography==44.0.2
ecdsa==0.19.1
et_xmlfile==2.0.0
ezodf==0.3.2
Flask==3.1.0
flask-cors==5.0.1
Flask-SQLAlchemy==3.1.1
google-api-core==2.28.1
google-api-python-client==2.178.0
google-auth==2.40.1
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
google-cloud-core==2.5.0
google-cloud-storage==3.6.0
google-crc32c==1.7.1
google-resumable-media==2.8.0
googleapis-common-protos==1.70.0
gspread==6.2.1
gunicorn==23.0.0
httplib2==0.22.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
lxml==5.4.0
MarkupSafe==3.0.2
numpy==2.2.6
oauthlib==3.2.2
openpyxl==3.1.5
packaging==25.0
pandas==2.2.3
pg8000==1.31.2
proto-plus==1.26.1
protobuf==6.31.0
psycopg2-binary==2.9.10
pyasn1==0.4.8
pyasn1_modules==0.4.1
pycparser==2.22
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-jose==3.4.0
pytz==2025.2
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9.1
scramp==1.4.5
six==1.17.0
SQLAlchemy==2.0.40
stripe==12.5.0
typing_extensions==4.13.2
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.4.0
Werkzeug==3.1.3
